requirements.txt
Django>=5.2.7
djangorestframework>=3.14.0
psycopg>=3.1.8
requests-html>=0.10.0
lxml-html-clean>=0.4.3
dateparser>=1.2.
gunicorn>=21.2.0